Cleanup & fix velocity
