+++ src/core/foundation.hpp
@@ -1336,11 +1336,6 @@ struct VariableResolver {
 };
 
 template <typename T> T *varAsObjectChecked(const SHVar &var, const shards::Type &type) {
-  // auto getObjectTypeName = [](SHTypeInfo type) {
-  //     auto objectInfo = shards::findObjectInfo(var.payload.objectVendorId, var.payload.objectTypeId);
-  //     auto name = objectInfo ? objectInfo->name : "<unknown>";
-  // };
-
   SHTypeInfo typeInfo(type);
   if (var.valueType != SHType::Object) {
     SHLOG_FATAL("Invalid type, expected: {} got: {}", type, magic_enum::enum_name(var.valueType));
